LayoutTests/platform/win/imported/w3c/web-platform-tests/cookies/partitioned-cookies/partitioned-cookies-samesite-attribute.https-expected.txt



PASS In top-level contexts, partitioned cookies default to the same SameSite attribute as unpartitioned cookies.
FAIL In top-level contexts, partitioned cookies can be set with all SameSite attributes. assert_equals: expected "Strict" but got "None"
